Refactor OAuth2Server into ResourceServer and AuthorizationServer
